buildtools/burg/table.c(52): remark #2259: non-pointer conversion from "NonTerminalNum={int}" to "short" may lose significant bits
http://monetdb.cwi.nl/testing/projects/monetdb/testweb/logs/39245:124139f4ff4a/Int-Fedora-i386-propcheck/make.html#l669

+++ buildtools/burg/table.c
@@ -49,7 +49,7 @@ addRelevant(r, nt) Relevant r; NonTerminalNum nt;
 		}
 	}
 	if (!r[i]) {
-		r[i] = nt;
+		r[i] = (short)nt;
 	}
 }
